Fix bug in useVirtualList hook
Handle state for visible items indepentently from other list
information. This makes it possible to render containers with correct
heights and offsets __before__ calculating number of visible items and
their offsets, and thus to avoid a race problem that occurs when the
list data changes but containers heights and offsets still are derived
from previous data.