Fix bug introduced in previous commit
In case the GalleryView component is not unmounted
between gallery change, the RTK query used for
fetching article data may hold true for both
isSuccess and isFetching at the same time, since
previous query succeeded and a new one is in
flight. This also means that currenData is
undefined whenever that happens, hence this minor
fix at the moment.