Add client-side based sort mechanism for media
By default mediasets is sorted by creation date
on the server. This commit adds the possibility
to extend this behavior by enabling sorting on the
client-side for entire sets of media (sets that
has not been sliced on the server) and includes,
as a starter, such an option for title in the
GalleryView.